Upgrade to Jest 2.4.7+jackson (#4107)
(cherry picked from commit fa86b1228e554be57dae39aee8a35306b5c6060e)

+++ graylog2-server/src/main/java/org/graylog2/indexer/searches/Searches.java
@@ -283,7 +283,7 @@ public class Searches {
         final long tookMs = tookMsFromSearchResult(searchResult);
         recordEsMetrics(tookMs, config.range());
 
-        return new SearchResult(hits, searchResult.getTotal().longValue(), indexRanges, config.query(), requestBuilder.toString(), tookMs);
+        return new SearchResult(hits, searchResult.getTotal(), indexRanges, config.query(), requestBuilder.toString(), tookMs);
     }
 
     private long tookMsFromSearchResult(io.searchbox.core.SearchResult searchResult) {
